# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.BatchRetryRequest:
    properties:
      limit:
        example: 100
        type: integer
      status:
        example: fail
        type: string
      task_ids:
        items:
          type: string
        type: array
      worker_name:
        example: my-worker
        type: string
    type: object
  dto.BatchRetryResponse:
    properties:
      new_task_ids:
        items:
          type: string
        type: array
      status:
        example: ok
        type: string
      total_retried:
        example: 10
        type: integer
    type: object
...
  dto.CreateTaskRequest:
    type: object
  dto.CreateTaskResponse:
    properties:
      asynq_task_id:
        type: string
      queue:
        example: default
        type: string
      status:
        example: pending
        type: string
      task_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      worker_name:
        example: my-worker
        type: string
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        example: 错误信息
        type: string
    type: object
...
  dto.ReplayTaskRequest:
    properties:
      delay:
        example: 0
        type: integer
    type: object
  dto.ReplayTaskResponse:
    properties:
      new_task_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      status:
        example: ok
        type: string
    type: object
  dto.ReportAttemptRequest:
    type: object
  dto.SuccessResponse:
    properties:
      data: {}
      message:
        example: 操作成功
        type: string
      status:
        example: ok
        type: string
    type: object
  dto.TaskListResponse:
    properties:
      items: {}
      total:
        type: integer
    type: object
  dto.TaskResponse:
    properties:
      task: {}
    type: object
...
host: localhost:28080
info:
  contact:
    name: TaskPM Support
  description: 分布式任务管理系统 - 基于 Asynq 和 PostgreSQL 的任务调度平台
  license:
    name: MIT
  title: TaskPM API
  version: 1.0.0
paths:
...
  /tasks:
    get:
      description: 分页查询任务列表，支持多条件过滤
      parameters:
      - description: Worker 名称
        in: query
        name: worker_name
        type: string
      - description: 任务状态
        in: query
        name: status
        type: string
      - description: 队列名称
        in: query
        name: queue
        type: string
      - default: 50
        description: 每页数量
        in: query
        name: limit
        type: integer
      - default: 0
        description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TaskListResponse'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 查询任务列表
      tags:
      - Tasks
    post:
      consumes:
      - application/json
      description: 创建新的异步任务并入队到 Asynq
      parameters:
      - description: 任务创建请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CreateTaskResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 创建任务
      tags:
      - Tasks
  /tasks/{task_id}:
    get:
      description: 根据 task_id 获取任务详细信息及执行历史
      parameters:
      - description: 任务 ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TaskResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 获取任务详情
      tags:
      - Tasks
  /tasks/{task_id}/replay:
    post:
      consumes:
      - application/json
      description: 将已完成或失败的任务重新入队执行
      parameters:
      - description: 任务 ID
        in: path
        name: task_id
        required: true
        type: string
      - description: 重放参数
        in: body
        name: request
        schema:
          $ref: '#/definitions/dto.ReplayTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ReplayTaskResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 重放任务
      tags:
      - Tasks
  /tasks/{task_id}/report-attempt:
    post:
      consumes:
      - application/json
      description: Worker 上报任务执行的详细状态（running、success、fail）
      parameters:
      - description: 任务 ID
        in: path
        name: task_id
        required: true
        type: string
      - description: 执行状态
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ReportAttemptRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 上报任务执行状态
      tags:
      - Tasks
  /tasks/batch-retry:
    post:
      consumes:
      - application/json
      description: 批量重试指定条件的失败任务
      parameters:
      - description: 重试条件
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.BatchRetryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BatchRetryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 批量重试失败任务
      tags:
      - Tasks
schemes:
- http
- https
swagger: "2.0"